feat: PendingRequestDTO id/createTime改為Long

+++ backend2/src/main/java/com/cmoney/backend2/forumocean/service/api/group/v2/PendingRequestDTO.kt
@@ -7,9 +7,9 @@ import com.google.gson.annotations.SerializedName
 @Keep
 data class PendingRequestDTO(
     @SerializedName("createTime")
-    val createTime: Int?,
+    val createTime: Long?,
     @SerializedName("id")
-    val id: Int?,
+    val id: Long?,
     @SerializedName("reason")
     val reason: String?
 )
